creator/requirements.txt
# Creator Bot - Requirements
# Python 3.10+

discord.py>=2.3.2
psycopg2-binary>=2.9.9
aiohttp>=3.9.1
beautifulsoup4>=4.12.2
lxml>=4.9.3
python-dotenv>=1.0.0
